mission 9 pt 1 vid 7

+++ Mission09_jrsarge/Models/Book.cs
@@ -12,14 +12,14 @@ namespace Mission09_jrsarge.Models
     {
         [Key]
         [Required]
-        public long BookId { get; set; }
+        public int BookId { get; set; }
         public string Title { get; set; }
         public string Author { get; set; }
         public string Publisher { get; set; }
         public string Isbn { get; set; }
         public string Classification { get; set; }
         public string Category { get; set; }
-        public long PageCount { get; set; }
+        public int PageCount { get; set; }
         public double Price { get; set; }
     }
 }
